refac(zk): move `zk/plonk/prover/*` to `zk/plonk/halo2/`
