+++ tachyon/zk/plonk/halo2/BUILD.bazel
@@ -3,6 +3,17 @@ load("//bazel:tachyon_cc.bzl", "tachyon_cc_library", "tachyon_cc_unittest")
 package(default_visibility = ["//visibility:public"])
 
 tachyon_cc_library(
+    name = "argument",
+    hdrs = ["argument.h"],
+    deps = [
+        "//tachyon/base:logging",
+        "//tachyon/base/containers:container_util",
+        "//tachyon/zk/base/entities:prover_base",
+        "//tachyon/zk/plonk/circuit:ref_table",
+    ],
+)
+
+tachyon_cc_library(
     name = "blake2b_transcript",
     hdrs = ["blake2b_transcript.h"],
     deps = [
@@ -160,6 +171,17 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
+    name = "synthesizer",
+    hdrs = ["synthesizer.h"],
+    deps = [
+        ":witness_collection",
+        "//tachyon/base/containers:container_util",
+        "//tachyon/zk/base/entities:prover_base",
+        "//tachyon/zk/plonk:constraint_system",
+    ],
+)
+
+tachyon_cc_library(
     name = "verifier",
     hdrs = ["verifier.h"],
     deps = [
@@ -175,21 +197,40 @@ tachyon_cc_library(
     ],
 )
 
+tachyon_cc_library(
+    name = "witness_collection",
+    hdrs = ["witness_collection.h"],
+    deps = [
+        "//tachyon/base:range",
+        "//tachyon/base/containers:container_util",
+        "//tachyon/zk/plonk/circuit:assignment",
+        "//tachyon/zk/plonk/circuit:phase",
+        "@com_google_absl//absl/container:btree",
+    ],
+)
+
 tachyon_cc_unittest(
     name = "halo2_unittests",
     srcs = [
+        "argument_unittest.cc",
         "blake2b_transcript_unittest.cc",
         "pinned_verifying_key_unittest.cc",
         "poseidon_transcript_unittest.cc",
         "proof_serializer_unittest.cc",
         "random_field_generator_unittest.cc",
         "sha256_transcript_unittest.cc",
+        "synthesizer_unittest.cc",
+        "witness_collection_unittest.cc",
     ],
     deps = [
+        ":argument",
         ":blake2b_transcript",
         ":pinned_verifying_key",
         ":poseidon_transcript",
         ":prover_test",
         ":sha256_transcript",
+        ":synthesizer",
+        "//tachyon/zk/plonk/circuit/examples:simple_circuit",
+        "//tachyon/zk/plonk/circuit/floor_planner:simple_floor_planner",
     ],
 )
