--- tachyon/zk/plonk/prover/BUILD.bazel
@@ -1,55 +1,0 @@
-load("//bazel:tachyon_cc.bzl", "tachyon_cc_library", "tachyon_cc_unittest")
-
-package(default_visibility = ["//visibility:public"])
-
-tachyon_cc_library(
-    name = "argument",
-    hdrs = ["argument.h"],
-    deps = [
-        "//tachyon/base:logging",
-        "//tachyon/base/containers:container_util",
-        "//tachyon/zk/base/entities:prover_base",
-        "//tachyon/zk/plonk/circuit:ref_table",
-    ],
-)
-
-tachyon_cc_library(
-    name = "synthesizer",
-    hdrs = ["synthesizer.h"],
-    deps = [
-        ":witness_collection",
-        "//tachyon/base/containers:container_util",
-        "//tachyon/zk/base/entities:prover_base",
-        "//tachyon/zk/plonk:constraint_system",
-    ],
-)
-
-tachyon_cc_library(
-    name = "witness_collection",
-    hdrs = ["witness_collection.h"],
-    deps = [
-        "//tachyon/base:range",
-        "//tachyon/base/containers:container_util",
-        "//tachyon/zk/plonk/circuit:assignment",
-        "//tachyon/zk/plonk/circuit:phase",
-        "@com_google_absl//absl/container:btree",
-    ],
-)
-
-tachyon_cc_unittest(
-    name = "prover_unittests",
-    srcs = [
-        "argument_unittest.cc",
-        "synthesizer_unittest.cc",
-        "witness_collection_unittest.cc",
-    ],
-    deps = [
-        ":argument",
-        ":synthesizer",
-        "//tachyon/base/containers:container_util",
-        "//tachyon/zk/plonk/circuit/examples:simple_circuit",
-        "//tachyon/zk/plonk/circuit/floor_planner:simple_floor_planner",
-        "//tachyon/zk/plonk/halo2:pinned_verifying_key",
-        "//tachyon/zk/plonk/halo2:prover_test",
-    ],
-)
